feat: Trim Entity 구현
